+++ src/apps/backend/modules/notification/internals/account_notification_preferences_reader.py
@@ -2,7 +2,7 @@ from typing import Optional
 from modules.notification.internals.store.account_notification_preferences_repository import (
     AccountNotificationPreferencesRepository,
 )
-from modules.account.internal.account_notification_preferences_util import AccountNotificationPreferenceUtil
+from modules.notification.internals.account_notification_preferences_util import AccountNotificationPreferenceUtil
 from modules.notification.errors import NotificationPreferencesNotFoundError
 from modules.notification.types import NotificationPreferencesParams
