+++ src/apps/backend/modules/notification/internals/account_notification_preferences_writer.py
@@ -5,7 +5,7 @@ from modules.notification.internals.store.account_notification_preferences_model
 from modules.notification.internals.store.account_notification_preferences_repository import (
     AccountNotificationPreferencesRepository,
 )
-from modules.account.internal.account_notification_preferences_util import AccountNotificationPreferenceUtil
+from modules.notification.internals.account_notification_preferences_util import AccountNotificationPreferenceUtil
 from modules.notification.types import NotificationPreferencesParams
 
 
@@ -33,7 +33,7 @@ class AccountNotificationPreferenceWriter:
     def create_or_update_notification_preferences(
         account_id: str, preferences: NotificationPreferencesParams
     ) -> NotificationPreferencesParams:
-        from modules.account.internal.account_notification_preferences_reader import AccountNotificationPreferenceReader
+        from modules.notification.internals.account_notification_preferences_reader import AccountNotificationPreferenceReader
 
         existing_preferences = AccountNotificationPreferenceReader.get_existing_notification_preferences_by_account_id(
             account_id
